+++ src/components/TextFieldHook/TextFieldHook.scss
@@ -1,0 +1,24 @@
+@import "../../styles/ColorSchema.scss";
+
+.input {
+  width: 95%;
+  height: 35px;
+  padding-left: 15px;
+  border: none;
+  border-radius: 5px;
+  margin-bottom: 20px;
+  background: grey;
+  font-family: $zapFont;
+}
+label {
+  font-weight: 100;
+  text-align: left;
+  margin-left: 13px;
+  font-family: $zapFont;
+  text-transform: uppercase;
+  font-size: 13px;
+  padding-left: 15px;
+  padding-bottom: 2px;
+  color: black;
+  display: block;
+}
