+++ src/stories/grid.stories.js
@@ -7,13 +7,13 @@
 // storiesOf("Grid", module).add("with h1", () => (
 //   <div>
 //     <Grid nBox={4} m={3}>
-//       nBox={4} m={3}{" "}
+//       {/* nBox={4} m={3}{" "} */} <h1>grid</h1>
 //     </Grid>
-//     <Grid nBox={3} m={3}>
+//     {/* <Grid nBox={3} m={3}>
 //       nBox={3} m={3}{" "}
 //     </Grid>
 //     <Grid nBox={5} m={3}>
 //       nBox={5} m={3}{" "}
-//     </Grid>
+//     </Grid> */}
 //   </div>
 // ));
